Photo service corrections
