+++ PickAPod/PickAPod.Photo/Program.cs
@@ -1,3 +1,5 @@
+using PickAPod.Photo.Services;
+
 var builder = WebApplication.CreateBuilder(args);
 builder.Services.AddControllers();
 builder.Services.AddEndpointsApiExplorer();
